refactor(memory): split the autograph queue/worker out of service.rs
service.rs had grown to 1831 lines, well past the crate's 500-NLOC-per-file
budget it otherwise enforces (fused_recall.rs and reinforce.rs were already
split out for the same reason). Extract the deferred-autograph subsystem —
AutographJob, AutographQueue, AutographWorkerHandle, autograph_if,
autograph_dropped, autograph_queue_open, has_autograph, autograph,
fact_exists and spawn_autograph_worker — into a new autograph.rs child
module, following the existing #[path = "..."] pattern so it keeps full
access to MemoryService's private fields and methods.

spawn_autograph_worker was itself 64 NLOC, over the crate's 50-NLOC-per-
function budget — previously unflagged only because it was buried inside
the oversized file. Split it into three: spawn_autograph_worker (the
orchestrator), install_autograph_sender (the guarded sender install), and
drain_autograph_queue (the worker thread's loop body). No behavior change.

AutographWorkerHandle stays re-exported at its existing crate::service::
AutographWorkerHandle path. service.rs drops to 1537 lines.

Verified: cargo test -p velesdb-memory --features persistence
(671 passed, 0 failed), clippy --all-targets -D warnings -D
clippy::pedantic clean, cargo check -p velesdb-wasm --no-default-features
--target wasm32-unknown-unknown clean, and a local lizard run confirms no
function in the new file exceeds NLOC 50 / CCN 8.
